Fix two classes in same file
